Separate addTodoFom to new file
